move pricing card component to ui.tsx
