+++ src/components/ui.tsx
@@ -2,8 +2,9 @@
 // hand drawn arrows from svgrepo.com
 import NextLink from "next/link"
 import { twMerge } from "tailwind-merge"
-import { ComponentProps } from "react"
+import { ComponentProps, ReactNode } from "react"
 import { Show } from "@/lib/util/react"
+import { Card } from "flowbite-react"
 
 export const buttonStyling =
   "text-white bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-primary-600 dark:hover:bg-primary-700 focus:outline-none dark:focus:ring-primary-800"
@@ -584,3 +585,64 @@ export const MissingData = ({
     </div>
   </div>
 )
+
+const freeFeatures = ["Individual configuration", "No setup, or hidden fees", "3 free receipts"]
+const freeNonFeatures = ["Unlimited receipts", "Automatic emailing"]
+const paidFeatures = [
+  "Individual configuration",
+  "No setup, or hidden fees",
+  "Unlimited receipts",
+  "Automatic emailing",
+]
+export function PricingCard({
+  plan,
+  title: propsTitle,
+  button,
+}: {
+  plan: "pro" | "free"
+  title?: string
+  button?: ReactNode
+}) {
+  const isPro = plan === "pro"
+  const features = isPro ? paidFeatures : freeFeatures
+  const price = isPro ? 20 : 0
+
+  const title = propsTitle ?? (isPro ? "Pro Plan" : "Free Plan")
+
+  return (
+    <Card>
+      <h5 className="mb-4 text-xl font-medium text-gray-500 dark:text-gray-400">{title}</h5>
+      <div className="flex items-baseline text-gray-900 dark:text-white">
+        <span className="text-3xl font-semibold">CAD</span>
+        <span className="text-5xl font-extrabold tracking-tight">{price}</span>
+        <span className="ml-1 text-xl font-normal text-gray-500 dark:text-gray-400">/year</span>
+      </div>
+      <ul className="my-7 space-y-5">
+        {features.map((feature, idx) => (
+          <li
+            key={idx}
+            className="flex space-x-3 text-base font-normal leading-tight text-gray-500 dark:text-gray-400"
+          >
+            <div className="-mb-1 mr-4 inline-block w-5 text-green-300">
+              <Svg.Tick />
+            </div>
+            {feature}
+          </li>
+        ))}
+        {!isPro &&
+          freeNonFeatures.map((nonFeature, idx) => (
+            <li
+              key={idx}
+              className="flex space-x-3 text-base font-normal leading-tight text-gray-500 line-through decoration-gray-500"
+            >
+              <div className="-mb-1 mr-4 inline-block w-5 text-red-300">
+                <Svg.Cross />
+              </div>
+              {nonFeature}
+            </li>
+          ))}
+      </ul>
+      {button}
+    </Card>
+  )
+}
